% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_indicators.R
\name{get_indicators}
\alias{get_indicators}
\title{get population measurements (indicators) from the UN population API.}
\usage{
get_indicators(theIndicators)
}
\arguments{
\item{theIndicators}{A string}
}
\value{
A data frame'
}
\description{
get population measurements (indicators) from the UN population API.
}
\examples{
get_indicators(1)
get_indicators(paste(c(1,3),collapse=","))
}
