% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_locations.R
\name{get_locations}
\alias{get_locations}
\title{get the geographical areas (locations) included in the UN population API.}
\usage{
get_locations(theLocations)
}
\arguments{
\item{theLocations}{A string}
}
\value{
A data frame'
}
\description{
get the geographical areas (locations) included in the UN population API.
}
\examples{
get_locations(704)
get_locations("VN")
get_locations(paste(c(1,3),collapse=","))
}
